Replace TIle Tool now supports infinite maps
This was actually part of a forward-looking update for detailed selections. If SelectedArea is updated with a rects() method the same way TileLayer.regions() is going to be in 1.8, then this script should start using proper selections rather than just their bounding boxes. For now, it necessarily falls back to using the bounding box because there is no `rects()`.