Terrain Rectangle now uses tile probability, handles Mixed terrains better
When using a Mixed terrain set, the tool now checks whether the chosen colour is used as corners, rectangles, or both. If the current colour isn't used for both, then matching is treated as if it were a corner or edge set. This allows more styles of labelling to be usable by this tool. However, these newly usable Mixed terrains may have worse behaviour when the other labels are used for disambiguation, so this tool still works best with Corner and Edge sets.