Clarify locked layer behaviour on Tiled < 1.10
Prior to Tiled 1.10, the locked status of layers wasn't copied when layers were copied. When a layer is resized (such as when the map is resized), Tiled makes a copy of it, and this copy would no longer be locked in older versions of Tiled. Since the Insert Row/Col actions resize the map prior to checking layers (doing so early would leave us with a bunch of layers that no longer exist after the map is resized), we have no good way to know whether a layer should be locked or not.

As of Tiled 1.10, layer lock status is copied, resizing maps no longer unlocks all the Tile Layers, and all the spreadsheet actions work as expected.