Mass Replace Tiles: Flips, Projects, performance
The relative flips of the old and new tiles are now taken into account, allowing you to flip/rotate tiles as part of the replacement process. Tile Objects are not rotated, the diagonal flip flags are completely ignored for them.

In Tiled 1.10.1+, if a Project is loaded, an additional batch action will be made available, applying the mass replacer to all maps in the project.

Improved performance by caching the old-new tile mappings. This uses more memory, but is about 15x faster.

Raises minimum supported Tiled version to 1.8, using region.rects to avoid iterating empty regions of maps.