Add Select Child Layers script
I'd previously uploaded this to the Tiled Discord for someone else, but it's turned out surprisingly useful even to me, who does not use a lot of layers, so I figured I should share it here. I've also added the action to the Layers view right-click menu, so it's more convenient to use.