Automap Helper: Fix Convert Brush to Layers, fix whitespace
Before this fix to Convert Brush to Layers, when a single-layer brush was converted to multiple layers, this would leave an empty layer, which could cause issues - most notably, it would cause further attempts to convert the brush (e.g. to another set of layers) to fail. Now, this starter layer is deleted if it's empty.

The whitespace is now consistently tabs everywhere, instead of a mix of tabs and spaces.