Mass Replace Tiles: Prompt for remapper
Now, if you run "Mass Replace Tiles" on a regular map, the script will prompt you for the remapper to use from among valid remappers open in Tiled, if it finds any. This is much more convenient than setting the "remappingMap" custom property for one-off replacements. The "remappingMap" property takes priority if present.