Move Tiles Tool: Reduce undos, handle empty better
In Cut mode, there is now only one Undo entry even if cutting from multiple layers.

When moving the tiles, only non-empty tiles are included in the preview, and empty parts of the brush will no longer overwrite tiles when comitting the change. That was a rather silly oversight in the original code - I skipped empty tiles when building the brush, but not when copying empty tiles from the brush to the preview.