Mass Replace Tiles: Warn about additional layers
People occasionally try to use their current map as the remapper, aiming to remap just that map. This workflow is not currently supported (maybe one day!). For now, I've added clearer text to the instructions that a remapping map will not remap itself, and also added a note about this to the "would you like to use this remapping map?" prompt when non-remapping layers are detected.

Since this workflow seems to be expected, I would like to look into supporting it eventually, but for now, I don't have the energy. It would require some more substantial changes to how the remapping is done.