Tileset Heatmap: Skip maps that don't use tileset
There's no sense in scanning a map if it doesn't even contain a reference to the tileset we're looking for. This should hopefully speed up scanning a little.

The check is against `map.tilesets` rather than `map.usedTilesets()`, in the interest of avoiding scanning the map's contents twice. This check should be much faster when most tilesets are actually used, which I would hope is the more common scenario.