Tileset Heatmap: Allow skipping subdirectories
I had the scanSubfolders config boolean already there, but forgot to actually use it :] Fortunately I didn't document it before.

Also, changed where the ExcludeFromHeatmaps check happens. It's probably slightly less efficient now, but the code is easier to follow as the checks all happen in the same place, instead of being scattered throughout the script.